- post 수정, comment 수정 리팩토링

+++ src/main/java/com/project/mini/service/PostService.java
@@ -39,7 +39,9 @@ public class PostService {
 
         Map<String, String> imgResult = s3Service.uploadFile(multipartFile);
 
-        User user = userRepo.findByUsername(userDetails.getUsername()).get();
+        User user = userRepo.findByUsername(userDetails.getUsername()).orElseThrow(
+                () -> new IllegalArgumentException("유저가 없습니다.")
+        );
 
         //새로운 게시글 등록
         Post post = new Post(dto, user, imgResult);
@@ -51,7 +53,9 @@ public class PostService {
 
     //디테일 페이지 게시글 조회
     public PostDetailResponseDto getDetailPost(Long postid) {
-        Post post = postRepo.findByPostId(postid).get();
+        Post post = postRepo.findByPostId(postid).orElseThrow(
+                () -> new IllegalArgumentException("게시글이 없습니다.")
+        );
         List<CommentResponseDto> commentList = new ArrayList<>();
         List<Comment> allByPost_postId = commentRepository.findAllByPost_PostId(postid);
         for (Comment comment : allByPost_postId) {
@@ -64,7 +68,9 @@ public class PostService {
     //게시글 삭제
 
     public void deletePost(Long postId, UserDetailsImpl userDetails) {
-        Post post = postRepo.findByPostId(postId).get();
+        Post post = postRepo.findByPostId(postId).orElseThrow(
+                () -> new IllegalArgumentException("게시글이 없습니다.")
+        );
 
         //post를 작성한 유저와 로그인한 유저가 같은 사람이면 수정가능 아니면 예외발생
         validationCheck(post, userDetails);
@@ -82,46 +88,61 @@ public class PostService {
         //이미지의 수정은 기존에 있던 이미지의 삭제 후 다시등록으로
 
         //넘겨받은 수정하고 싶은 postId를 받아와서 조회
-        Post post = postRepo.findByPostId(postId).get();
-
+        Post post = postRepo.findByPostId(postId).orElseThrow(
+                () -> new IllegalArgumentException("게시글이 없습니다.")
+        );
         //post를 작성한 유저와 로그인한 유저가 같은 사람이면 수정가능 아니면 예외발생
         validationCheck(post, userDetails);
-
-        //기존 이미지 삭제후 재등록
-        s3Service.deleteFile(post.getTransImgFileName());
-        Map<String, String> imgResult = s3Service.uploadFile(multipartFile);
-
-        //해피포인트 수정
+        //해피포인트 수정//해피포인트 수정
         post.getUser().modifyHappypoint(post.getHappypoint(), dto.getHappypoint());
+        if (multipartFile != null) {
+            //기존 이미지 삭제후 재등록
+            s3Service.deleteFile(post.getTransImgFileName());
+            Map<String, String> imgResult = s3Service.uploadFile(multipartFile);
 
-        //엔티티 업데이트
-        post.Update(dto, imgResult);
-
+            //엔티티 업데이트
+            post.Update(dto, imgResult);
+        } else {
+            //엔티티 업데이트
+            post.Update(dto);
+        }
         postRepo.save(post);
+
+//        //기존 이미지 삭제후 재등록
+//        s3Service.deleteFile(post.getTransImgFileName());
+//        Map<String, String> imgResult = s3Service.uploadFile(multipartFile);
+
+//        //해피포인트 수정
+//        post.getUser().modifyHappypoint(post.getHappypoint(), dto.getHappypoint());
+//
+//        //엔티티 업데이트
+//        post.Update(dto, imgResult);
+//
+//        postRepo.save(post);
     }
+
+//    public void modifyPost(Long postId, PostDto dto, UserDetailsImpl userDetails) {
+//        //이미지의 수정은 기존에 있던 이미지의 삭제 후 다시등록으로
+//
+//        //넘겨받은 수정하고 싶은 postId를 받아와서 조회
+//        Post post = postRepo.findByPostId(postId).get();
+//
+//        //post를 작성한 유저와 로그인한 유저가 같은 사람이면 수정가능 아니면 예외발생
+//        validationCheck(post, userDetails);
+//
+//        //해피포인트 수정
+//        post.getUser().modifyHappypoint(post.getHappypoint(), dto.getHappypoint());
+//
+//        //엔티티 업데이트
+//        post.Update(dto);
+//
+//        postRepo.save(post);
+//    }
 
     //게시글 작성자가 로그인한 유저인지확인
     private void validationCheck(Post post, UserDetailsImpl userDetails) {
         if (!post.getUser().getUsername().equals(userDetails.getUsername())) {
             throw new IllegalArgumentException("게시글 작성자만 조작할 수 있습니다.");
         }
-    }
-
-    public void modifyPost(Long postId, PostDto dto, UserDetailsImpl userDetails) {
-        //이미지의 수정은 기존에 있던 이미지의 삭제 후 다시등록으로
-
-        //넘겨받은 수정하고 싶은 postId를 받아와서 조회
-        Post post = postRepo.findByPostId(postId).get();
-
-        //post를 작성한 유저와 로그인한 유저가 같은 사람이면 수정가능 아니면 예외발생
-        validationCheck(post, userDetails);
-
-        //해피포인트 수정
-        post.getUser().modifyHappypoint(post.getHappypoint(), dto.getHappypoint());
-
-        //엔티티 업데이트
-        post.Update(dto);
-
-        postRepo.save(post);
     }
 }
